+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shopify/cli-hydrogen
 
+## 8.0.3
+
+### Patch Changes
+
+- Update internal `@shopify/cli-kit` dependency to fix React version mismatches. ([#2059](https://github.com/Shopify/hydrogen/pull/2059)) by [@frandiox](https://github.com/frandiox)
+
+## 8.0.2
+
+### Patch Changes
+
+- Pin React dependency to 18.2.0 to avoid mismatches. ([#2051](https://github.com/Shopify/hydrogen/pull/2051)) by [@frandiox](https://github.com/frandiox)
+
 ## 8.0.1
 
 ### Patch Changes